consolidate backlog tests
