+++ chain-signatures/node/src/backlog/mod.rs
@@ -699,1008 +699,4 @@ impl BacklogEntry {
 }
 
 #[cfg(test)]
-mod tests {
-    use crate::backlog::checkpoints::Checkpoint;
-    use crate::backlog::mock::{
-        bidi_initial_status, mock_bidi_request, mock_bidi_response, mock_bidi_response_request,
-        mock_bidirectional_tx, mock_execution_entry, mock_execution_entry_with_timestamp,
-        mock_participants, mock_publish_state, mock_publish_state_with_proposer, mock_sign_request,
-        mock_signature_output, mock_tx, pending_execution_status, single_entry_checkpoint,
-        BacklogTestExt,
-    };
-    use crate::backlog::{
-        AnyProgress, Backlog, BacklogEntry, BacklogError, Bidirectional, Executing, Final,
-        Generating, Initial, PendingRequests, Publishing,
-    };
-    use crate::sign_bidirectional::{BidirectionalProgress, SignProgress, SignStatus};
-    use mpc_chain_integration_core::StateManager;
-    use mpc_primitives::{Chain, ChainConfig as _, SignId, SignKind};
-    use std::convert::TryInto;
-    use std::sync::Arc;
-
-    fn digest_hex(hex_str: &str) -> [u8; 32] {
-        hex::decode(hex_str)
-            .unwrap()
-            .try_into()
-            .expect("digest hex must be 32 bytes")
-    }
-
-    #[tokio::test]
-    async fn test_backlog_chain_isolation() {
-        let backlog = Backlog::new();
-
-        let sign_id_eth = SignId::from_u8(1);
-        let sign_id_sol = SignId::from_u8(2);
-        let sign_id_near = SignId::from_u8(3);
-
-        backlog
-            .insert_mock_bidirectional(sign_id_eth, Chain::Ethereum)
-            .await;
-        backlog
-            .insert_mock_bidirectional(sign_id_sol, Chain::Solana)
-            .await;
-        backlog
-            .insert_mock_bidirectional(sign_id_near, Chain::NEAR)
-            .await;
-
-        // Verify correct transactions in each chain
-        assert!(backlog.get(Chain::Ethereum, &sign_id_eth).await.is_some());
-        assert!(backlog.get(Chain::Ethereum, &sign_id_sol).await.is_none());
-        assert!(backlog.get(Chain::Solana, &sign_id_sol).await.is_some());
-        assert!(backlog.get(Chain::Solana, &sign_id_eth).await.is_none());
-        assert!(backlog.get(Chain::NEAR, &sign_id_near).await.is_some());
-        assert!(backlog.get(Chain::NEAR, &sign_id_eth).await.is_none());
-    }
-
-    #[tokio::test]
-    async fn test_backlog_filter_by_status() {
-        let backlog = Backlog::new();
-
-        // Add transactions with different statuses to Ethereum
-        let sign_id0 = SignId::from_u8(0);
-        let sign_id1 = SignId::from_u8(1);
-        let tx2 = mock_bidirectional_tx(SignId::from_u8(2), Chain::Ethereum);
-        let tx3 = mock_bidirectional_tx(SignId::from_u8(3), Chain::Ethereum);
-
-        backlog
-            .insert_mock_bidirectional(sign_id1, Chain::Ethereum)
-            .await;
-        backlog.insert_mock_final(&tx2).await;
-        backlog.insert_mock_executing(&tx3).await;
-
-        // Add transactions to Solana
-        let tx4 = mock_tx(4);
-        backlog.insert_mock_executing(&tx4).await;
-
-        // Filter Ethereum by Pending execution
-        let eth_pending = backlog
-            .get_by::<Bidirectional<Executing>>(Chain::Ethereum, &tx3.sign_id())
-            .await;
-        assert!(eth_pending.is_some());
-
-        // Filter Ethereum by Initial Generating
-        let eth_awaiting = backlog
-            .get_by::<Bidirectional<Initial<Generating>>>(Chain::Ethereum, &sign_id1)
-            .await;
-        assert!(eth_awaiting.is_some());
-
-        // Filter Ethereum by bidirectional completion awaiting final respond
-        let eth_completion = backlog
-            .get_by::<Bidirectional<Final<Generating>>>(Chain::Ethereum, &tx2.sign_id())
-            .await;
-        assert!(eth_completion.is_some());
-
-        // Filter Solana by Pending execution
-        let sol_pending = backlog
-            .get_by::<Bidirectional<Executing>>(Chain::Solana, &tx4.sign_id())
-            .await;
-        assert!(sol_pending.is_some());
-
-        // Filter non-existent chain returns empty
-        let near_pending = backlog
-            .get_by::<Bidirectional<Executing>>(Chain::NEAR, &sign_id0)
-            .await;
-        assert!(near_pending.is_none());
-    }
-
-    #[tokio::test]
-    async fn test_backlog_concurrent_access() {
-        let backlog = Backlog::new();
-        let mut handles = vec![];
-
-        // Spawn multiple tasks that insert concurrently to different chains
-        for i in 0..5 {
-            let backlog = backlog.clone();
-            let handle = tokio::spawn(async move {
-                backlog
-                    .insert_mock_bidirectional(SignId::from_u8(i), Chain::Ethereum)
-                    .await;
-            });
-            handles.push(handle);
-        }
-
-        for i in 5..10 {
-            let backlog = backlog.clone();
-            let handle = tokio::spawn(async move {
-                backlog
-                    .insert_mock_bidirectional(SignId::from_u8(i), Chain::Solana)
-                    .await;
-            });
-            handles.push(handle);
-        }
-
-        // Wait for all insertions and verify all were inserted
-        for handle in handles {
-            handle.await.unwrap();
-        }
-        assert_eq!(backlog.len_by_chain(Chain::Ethereum).await, 5);
-        assert_eq!(backlog.len_by_chain(Chain::Solana).await, 5);
-
-        // Spawn multiple tasks that remove concurrently
-        let mut handles = vec![];
-        for i in 0..5 {
-            let backlog = backlog.clone();
-            let handle = tokio::spawn(async move {
-                let id = SignId::from_u8(i);
-                backlog.remove(Chain::Ethereum, &id).await
-            });
-            handles.push(handle);
-        }
-
-        // Wait for all removals
-        for handle in handles {
-            let removed = handle.await.unwrap();
-            assert!(removed.is_some());
-        }
-
-        // Verify Ethereum chain is now empty, but Solana still has data
-        assert_eq!(backlog.len_by_chain(Chain::Ethereum).await, 0);
-        assert_eq!(backlog.len_by_chain(Chain::Solana).await, 5);
-    }
-
-    #[tokio::test]
-    async fn test_checkpoint_creation() {
-        let backlog = Backlog::new();
-
-        // Add some transactions
-        let tx1 = mock_bidirectional_tx(SignId::from_u8(1), Chain::Ethereum);
-        let tx2 = mock_bidirectional_tx(SignId::from_u8(2), Chain::Ethereum);
-
-        backlog.insert_mock_executing(&tx1).await;
-        backlog.insert_mock_final(&tx2).await;
-
-        backlog.set_processed_block(Chain::Ethereum, 100).await;
-
-        let checkpoint = backlog.checkpoint(Chain::Ethereum).await.unwrap();
-
-        assert_eq!(checkpoint.block_height, 100);
-        assert_eq!(checkpoint.pending_requests.len(), 2);
-        // Guard the checkpoint digest wire format; update only for intentional changes.
-        assert_eq!(
-            checkpoint.digest(),
-            digest_hex("884b11ef5550724b788b7e29e9a07e7a6fd46f94d604e6d38bae71a36816b65e")
-        );
-    }
-
-    #[test]
-    fn test_checkpoint_equality() {
-        let tx1 = mock_tx(1);
-        let tx2 = mock_tx(2);
-
-        let mut pending1 = PendingRequests::new();
-        pending1.insert(
-            tx1.sign_id(),
-            mock_execution_entry(&tx1, Chain::Ethereum, bidi_initial_status()),
-        );
-        pending1.insert(
-            tx2.sign_id(),
-            mock_execution_entry(&tx2, Chain::Ethereum, bidi_initial_status()),
-        );
-        pending1.set_processed_block(100);
-
-        let mut pending2 = PendingRequests::new();
-        pending2.insert(
-            tx1.sign_id(),
-            mock_execution_entry(&tx1, Chain::Ethereum, bidi_initial_status()),
-        );
-        pending2.insert(
-            tx2.sign_id(),
-            mock_execution_entry(&tx2, Chain::Ethereum, bidi_initial_status()),
-        );
-        pending2.set_processed_block(100);
-
-        let checkpoint1 = Checkpoint::snapshot(&pending1, Chain::Ethereum);
-        let checkpoint2 = Checkpoint::snapshot(&pending2, Chain::Ethereum);
-        // Same data should be equal
-        assert_eq!(checkpoint1, checkpoint2);
-        assert_eq!(checkpoint1.digest(), checkpoint2.digest());
-
-        // Different block height should not be equal
-        let mut checkpoint3 = Checkpoint::snapshot(&pending2, Chain::Ethereum);
-        checkpoint3.block_height = 101;
-        assert_ne!(checkpoint1, checkpoint3);
-    }
-
-    #[test]
-    fn test_checkpoint_consensus_projection() {
-        let tx = mock_bidirectional_tx(SignId::from_u8(60), Chain::Ethereum);
-        let sign_id = tx.sign_id();
-
-        // The digest commits only to the consensus projection of each entry's
-        // status (sorted by sign_id), not to the request or publish content.
-
-        // Initial source-chain phase: generation, and publishing by any proposer,
-        // all collapse to a single digest.
-        let generation = single_entry_checkpoint(mock_execution_entry(
-            &tx,
-            Chain::Ethereum,
-            bidi_initial_status(),
-        ));
-        let publish = single_entry_checkpoint(mock_execution_entry(
-            &tx,
-            Chain::Ethereum,
-            SignStatus::Bidirectional(BidirectionalProgress::Initial(SignProgress::Publishing(
-                mock_publish_state(),
-            ))),
-        ));
-        let publish_other = single_entry_checkpoint(mock_execution_entry(
-            &tx,
-            Chain::Ethereum,
-            SignStatus::Bidirectional(BidirectionalProgress::Initial(SignProgress::Publishing(
-                mock_publish_state_with_proposer(false),
-            ))),
-        ));
-        assert_eq!(generation.digest(), publish.digest());
-        assert_eq!(generation.digest(), publish_other.digest());
-
-        // Plain `Sign` requests follow the same initial-phase projection.
-        let plain = mock_sign_request(sign_id, Chain::Ethereum);
-        let plain_generation = single_entry_checkpoint(BacklogEntry::new(Arc::clone(&plain)));
-        let plain_publish = single_entry_checkpoint(BacklogEntry::with_status(
-            plain,
-            SignStatus::Sign(SignProgress::Publishing(mock_publish_state())),
-        ));
-        assert_eq!(plain_generation.digest(), plain_publish.digest());
-
-        // Post-initial phase: awaiting target-chain execution and the final
-        // response generation/publish states are not observable at the
-        // source-chain checkpoint height, so they share the checkpoint digest.
-        let execution = single_entry_checkpoint(mock_execution_entry(
-            &tx,
-            Chain::Ethereum,
-            pending_execution_status(&tx),
-        ));
-        let response_request = mock_bidi_response(&tx);
-        let origin_request = mock_bidi_request(sign_id, Chain::Ethereum);
-        let gen_bidirectional = single_entry_checkpoint(BacklogEntry::with_status(
-            Arc::clone(&origin_request),
-            SignStatus::Bidirectional(BidirectionalProgress::Final {
-                respond_request: Arc::clone(&response_request),
-                progress: SignProgress::Generating,
-            }),
-        ));
-        let pub_bidirectional = single_entry_checkpoint(BacklogEntry::with_status(
-            origin_request,
-            SignStatus::Bidirectional(BidirectionalProgress::Final {
-                respond_request: response_request,
-                progress: SignProgress::Publishing(mock_publish_state()),
-            }),
-        ));
-        assert_eq!(
-            execution.digest(),
-            gen_bidirectional.digest(),
-            "PendingExecution must yield the same checkpoint digest as the final response generation state"
-        );
-        assert_eq!(
-            execution.digest(),
-            pub_bidirectional.digest(),
-            "PendingExecution must yield the same checkpoint digest as the final response publish state"
-        );
-
-        // The initial source-chain phase is observable at this height and must
-        // still differ from the post-initial phase.
-        assert_ne!(
-            generation.digest(),
-            execution.digest(),
-            "the initial source-chain phase must remain distinct in the checkpoint"
-        );
-    }
-
-    #[test]
-    fn test_respond_updates_entry_atomically() {
-        let tx = mock_bidirectional_tx(SignId::from_u8(23), Chain::Ethereum);
-        let mut entry = mock_execution_entry(&tx, Chain::Ethereum, pending_execution_status(&tx));
-        let response_request = mock_bidi_response(&tx);
-
-        entry.respond(response_request).unwrap();
-
-        assert_matches!(entry.request().kind, SignKind::RespondBidirectional(_));
-        assert_matches!(
-            entry.status(),
-            SignStatus::Bidirectional(BidirectionalProgress::Final {
-                progress: SignProgress::Generating,
-                ..
-            })
-        );
-    }
-
-    #[test]
-    fn test_respond_rejects_mismatched_request_id() {
-        let tx = mock_tx(24);
-        let original_sign_id = tx.sign_id();
-        let mut entry = mock_execution_entry(&tx, Chain::Ethereum, pending_execution_status(&tx));
-        let response_request =
-            mock_bidi_response_request(SignId::from_u8(25), tx.id, Chain::Ethereum);
-
-        let err = entry.respond(response_request).unwrap_err();
-
-        assert_matches!(err, BacklogError::InvalidBidirectionalResponseTransition);
-        assert_eq!(entry.sign_id(), original_sign_id);
-        assert_matches!(
-            entry.status(),
-            SignStatus::Bidirectional(BidirectionalProgress::Executing(_))
-        );
-    }
-
-    #[test]
-    fn test_checkpoint_digest_ignores_timestamp() {
-        let tx = mock_tx(8);
-
-        let entry1 =
-            mock_execution_entry_with_timestamp(&tx, Chain::Ethereum, bidi_initial_status(), 0);
-        let entry2 =
-            mock_execution_entry_with_timestamp(&tx, Chain::Ethereum, bidi_initial_status(), 9999);
-
-        let checkpoint1 = single_entry_checkpoint(entry1);
-        let checkpoint2 = single_entry_checkpoint(entry2);
-
-        assert_eq!(checkpoint1.digest(), checkpoint2.digest());
-    }
-
-    #[test]
-    fn test_checkpoint_digest_differs_for_different_requests() {
-        let tx1 = mock_tx(10);
-        let tx2 = mock_tx(11);
-
-        let checkpoint1 = single_entry_checkpoint(mock_execution_entry(
-            &tx1,
-            Chain::Ethereum,
-            bidi_initial_status(),
-        ));
-        let checkpoint2 = single_entry_checkpoint(mock_execution_entry(
-            &tx2,
-            Chain::Ethereum,
-            bidi_initial_status(),
-        ));
-
-        assert_ne!(checkpoint1.digest(), checkpoint2.digest());
-        assert_eq!(
-            checkpoint1.digest(),
-            digest_hex("a31e0d66f5b4fb860cc62e809cc29918b9138550b5cd62e1c752fc40ce6c2779")
-        );
-    }
-
-    #[test]
-    fn test_checkpoint_serialization() {
-        let tx1 = mock_tx(1);
-        let checkpoint = single_entry_checkpoint(mock_execution_entry(
-            &tx1,
-            Chain::Ethereum,
-            pending_execution_status(&tx1),
-        ));
-
-        // Test JSON serialization
-        let json = serde_json::to_string(&checkpoint).unwrap();
-        let deserialized: Checkpoint = serde_json::from_str(&json).unwrap();
-
-        assert_eq!(checkpoint, deserialized);
-        // Guard the checkpoint digest wire format; update only for intentional changes.
-        assert_eq!(
-            checkpoint.digest(),
-            digest_hex("12f5bc5c4f0fea1debafceb8879644ea545309775b3e2cc266335cd3247d5394")
-        );
-        assert_eq!(checkpoint.digest(), deserialized.digest());
-
-        let restored_entry = &deserialized.pending_requests[0];
-        assert_eq!(restored_entry.sign_id(), tx1.sign_id());
-        let SignKind::SignBidirectional(ref event) = restored_entry.request.kind else {
-            panic!("Expected SignBidirectional kind");
-        };
-        assert_eq!(event.dest, "test_dest");
-        assert_eq!(restored_entry.status, pending_execution_status(&tx1));
-    }
-
-    #[tokio::test]
-    async fn test_recover_restores_execution_watchers() {
-        let backlog = Backlog::new();
-        let tx = mock_tx(6);
-        let sign_id = tx.sign_id();
-
-        backlog.insert_mock_executing(&tx).await;
-        backlog.set_processed_block(Chain::Solana, 10).await;
-
-        let checkpoint = backlog.checkpoint(Chain::Solana).await.unwrap();
-
-        let recovered = Backlog::new();
-        recovered
-            .recover_by_checkpoint(checkpoint)
-            .await
-            .expect("failed to recover");
-
-        let entry = recovered
-            .get_by::<Bidirectional<Executing>>(Chain::Solana, &sign_id)
-            .await
-            .expect("entry should exist");
-        assert_eq!(entry.sign_id(), sign_id);
-        assert_eq!(&*entry.state().0 .0, &tx);
-
-        let watchers = recovered.get_execution_watchers(Chain::Ethereum).await;
-        assert_eq!(watchers.len(), 1);
-        assert!(watchers.contains_key(&tx.id));
-    }
-
-    #[tokio::test]
-    async fn test_recovery_makes_checkpoint_visible_as_latest() {
-        let backlog = Backlog::new();
-        let tx = mock_tx(16);
-
-        backlog.insert_mock_executing(&tx).await;
-        backlog.set_processed_block(Chain::Solana, 10).await;
-
-        let checkpoint = backlog.checkpoint(Chain::Solana).await.unwrap();
-
-        let recovered = Backlog::new();
-        recovered
-            .checkpoint_storage()
-            .persist(&checkpoint)
-            .await
-            .unwrap();
-        recovered
-            .recover_by_checkpoint(checkpoint.clone())
-            .await
-            .expect("failed to recover");
-
-        assert_eq!(
-            recovered.latest_checkpoint(Chain::Solana).await,
-            Some(checkpoint),
-            "recovered checkpoint should be visible via latest_checkpoint for /checkpoint"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_recover_preserves_sign_kind() {
-        let backlog = Backlog::new();
-        let sign_id = SignId::from_u8(42);
-
-        backlog
-            .insert_mock_bidirectional(sign_id, Chain::Solana)
-            .await;
-        backlog.set_processed_block(Chain::Solana, 10).await;
-
-        let checkpoint = backlog.checkpoint(Chain::Solana).await.unwrap();
-
-        let recovered = Backlog::new();
-        recovered
-            .recover_by_checkpoint(checkpoint)
-            .await
-            .expect("failed to recover");
-
-        let recovered_entry = recovered
-            .get(Chain::Solana, &sign_id)
-            .await
-            .expect("missing recovered entry");
-
-        assert_matches!(
-            recovered_entry.request().kind,
-            SignKind::SignBidirectional(_)
-        );
-    }
-
-    #[tokio::test]
-    async fn test_recovered_completed_bidirectional_requests_are_requeued_for_final_respond() {
-        for offset in 0..2 {
-            let backlog = Backlog::new();
-            let tx = mock_tx(8 + offset as u8);
-
-            backlog.insert_mock_final(&tx).await;
-            backlog.set_processed_block(Chain::Solana, 10).await;
-
-            let checkpoint = backlog.checkpoint(Chain::Solana).await.unwrap();
-
-            let recovered = Backlog::new();
-            recovered
-                .recover_by_checkpoint(checkpoint)
-                .await
-                .expect("failed to recover");
-
-            let requeued = recovered.take_requeueable_requests(Chain::Solana).await;
-            assert_eq!(
-                requeued.len(),
-                1,
-                "completed bidirectional request should be requeued for final respond"
-            );
-            assert_matches!(requeued[0].kind, SignKind::RespondBidirectional(_));
-        }
-    }
-
-    #[tokio::test]
-    async fn test_awaiting_response_bidirectional_requeues() {
-        let backlog = Backlog::new();
-        let tx = mock_tx(42);
-
-        backlog.insert_mock_final(&tx).await;
-
-        let requeued = backlog.take_requeueable_requests(Chain::Solana).await;
-        assert_eq!(requeued.len(), 1);
-        assert_matches!(requeued[0].kind, SignKind::RespondBidirectional(_));
-    }
-
-    #[tokio::test]
-    async fn test_publish_accepts_bidirectional_pending_generation() {
-        let backlog = Backlog::new();
-        let tx = mock_tx(43);
-        let sign_id = tx.sign_id();
-
-        let entry = backlog
-            .insert_mock_bidirectional(sign_id, Chain::Solana)
-            .await;
-
-        let (pk, output) = mock_signature_output(&entry.request().args);
-        let entry = entry
-            .advance(pk, &output, mock_participants(), true)
-            .await
-            .expect("pending generation should transition to publishing");
-
-        let fetched = backlog
-            .get_by::<Bidirectional<Initial<Publishing>>>(Chain::Solana, &sign_id)
-            .await
-            .expect("entry should remain in backlog");
-        assert_eq!(fetched.request().id, sign_id);
-        assert_eq!(entry.request().id, sign_id);
-    }
-
-    #[tokio::test]
-    async fn test_publish_accepts_final_respond_generation() {
-        let backlog = Backlog::new();
-        let tx = mock_tx(44);
-        let sign_id = tx.sign_id();
-
-        let entry = backlog.insert_mock_final(&tx).await;
-
-        let (pk, output) = mock_signature_output(&entry.request().args);
-        let entry = entry
-            .advance(pk, &output, mock_participants(), true)
-            .await
-            .expect("final publishing");
-
-        let fetched = backlog
-            .get_by::<Bidirectional<Final<Publishing>>>(Chain::Solana, &sign_id)
-            .await
-            .expect("entry should remain in backlog");
-        assert_eq!(fetched.request().id, sign_id);
-        assert_eq!(entry.request().id, sign_id);
-    }
-
-    #[tokio::test]
-    async fn test_watch_unwatch_and_respond() {
-        let backlog = Backlog::new();
-        let tx = mock_tx(7);
-        let sign_id = tx.sign_id();
-
-        // Insert an executing request on the source chain
-        backlog.insert_mock_executing(&tx).await;
-
-        // Watch execution on the target chain
-        backlog
-            .watch_execution(tx.target_chain, sign_id, Arc::new(tx.clone()))
-            .await;
-
-        // Unwatch should return the watcher
-        let maybe = backlog.unwatch_execution(tx.target_chain, &tx.id).await;
-        assert!(maybe.is_some());
-        let (s, watched_tx) = maybe.unwrap();
-        assert_eq!(s, sign_id);
-        assert_eq!(watched_tx.id, tx.id);
-
-        // respond should transition to final response signing
-        let completion_request = mock_bidi_response(&tx);
-        let entry = backlog
-            .get_by::<Bidirectional<Executing>>(tx.source_chain, &sign_id)
-            .await
-            .expect("executing entry should be found");
-        entry
-            .advance(completion_request)
-            .await
-            .expect("respond should transition to final generating");
-        assert!(backlog
-            .get_by::<Bidirectional<Final<Generating>>>(tx.source_chain, &sign_id)
-            .await
-            .is_some());
-    }
-
-    #[tokio::test]
-    async fn test_automatic_checkpoint_on_interval() {
-        let backlog = Backlog::new();
-
-        // Add some transactions
-        let sign_id1 = SignId::from_u8(1);
-        backlog
-            .insert_mock_bidirectional(sign_id1, Chain::Ethereum)
-            .await;
-
-        let interval = Chain::Ethereum.checkpoint_interval().unwrap();
-
-        // First few blocks shouldn't create checkpoints
-        for i in 1..interval {
-            let checkpoint = backlog.set_processed_block(Chain::Ethereum, i).await;
-            assert!(checkpoint.is_none(), "Block {i} should not make checkpoint");
-        }
-
-        // At block interval, should create checkpoint
-        let checkpoint = backlog.set_processed_block(Chain::Ethereum, interval).await;
-        assert!(checkpoint.is_some());
-        let checkpoint = checkpoint.unwrap();
-        assert_eq!(checkpoint.block_height, interval);
-        assert_eq!(checkpoint.chain, Chain::Ethereum);
-        assert_eq!(checkpoint.pending_requests.len(), 1);
-
-        let checkpoint = backlog
-            .set_processed_block(Chain::Ethereum, interval + 1)
-            .await;
-        assert!(checkpoint.is_none());
-
-        let checkpoint = backlog
-            .set_processed_block(Chain::Ethereum, 2 * interval)
-            .await;
-        assert!(checkpoint.is_some());
-        let checkpoint = checkpoint.unwrap();
-        assert_eq!(checkpoint.block_height, 2 * interval);
-    }
-
-    #[tokio::test]
-    async fn test_automatic_checkpoint_solana_interval() {
-        let backlog = Backlog::new();
-        let interval = Chain::Solana.checkpoint_interval().unwrap();
-
-        // Add transaction
-        let sign_id1 = SignId::from_u8(1);
-        backlog
-            .insert_mock_bidirectional(sign_id1, Chain::Solana)
-            .await;
-
-        // Solana interval is 10 blocks
-        for i in 1..interval {
-            let checkpoint = backlog.set_processed_block(Chain::Solana, i).await;
-            assert!(checkpoint.is_none(), "Block {i} should not make checkpoint");
-        }
-
-        // At block interval, should create checkpoint
-        let checkpoint = backlog.set_processed_block(Chain::Solana, interval).await;
-        assert!(checkpoint.is_some());
-        let checkpoint = checkpoint.unwrap();
-        assert_eq!(checkpoint.block_height, interval);
-        assert_eq!(checkpoint.chain, Chain::Solana);
-    }
-
-    async fn seed_pending_solana_request(backlog: &Backlog) {
-        backlog
-            .insert_mock_bidirectional(SignId::from_u8(1), Chain::Solana)
-            .await;
-    }
-
-    #[tokio::test]
-    async fn test_boundary_crossing_sparse_request_waits_until_next_bucket() {
-        let backlog = Backlog::new();
-        seed_pending_solana_request(&backlog).await;
-
-        // This documents the main caveat of boundary-crossing checkpointing:
-        // sparse requests still wait if the next observed slot remains in the
-        // same interval bucket.
-        //
-        // 480 -> 500: same bucket (both / 120 == 4), no checkpoint.
-        backlog
-            .set_processed_block_interval(Chain::Solana, 480, 120)
-            .await;
-        let cp = backlog
-            .set_processed_block_interval(Chain::Solana, 500, 120)
-            .await;
-        assert!(cp.is_none());
-
-        // 500 -> 600: crosses from bucket 4 to bucket 5
-        let cp = backlog
-            .set_processed_block_interval(Chain::Solana, 600, 120)
-            .await;
-        assert!(cp.is_some());
-        let cp = cp.unwrap();
-        assert_eq!(cp.block_height, 600);
-        assert_eq!(cp.pending_requests.len(), 1);
-    }
-
-    #[tokio::test]
-    async fn test_boundary_crossing_same_bucket_no_checkpoint() {
-        let backlog = Backlog::new();
-        seed_pending_solana_request(&backlog).await;
-
-        // 121 crosses from bucket 0 to bucket 1
-        let cp = backlog
-            .set_processed_block_interval(Chain::Solana, 121, 120)
-            .await;
-        assert!(cp.is_some());
-
-        // 130 stays in bucket 1; no new boundary crossed
-        let cp = backlog
-            .set_processed_block_interval(Chain::Solana, 130, 120)
-            .await;
-        assert!(cp.is_none());
-    }
-
-    #[tokio::test]
-    async fn test_boundary_crossing_within_first_bucket_no_checkpoint() {
-        let backlog = Backlog::new();
-        seed_pending_solana_request(&backlog).await;
-
-        // First observed slot 50 is still in bucket 0 (50 / 120 == 0 == prev default 0)
-        let cp = backlog
-            .set_processed_block_interval(Chain::Solana, 50, 120)
-            .await;
-        assert!(cp.is_none());
-    }
-
-    #[tokio::test]
-    async fn test_boundary_crossing_exact_multiple_still_checkpoints() {
-        let backlog = Backlog::new();
-        seed_pending_solana_request(&backlog).await;
-
-        let cp = backlog
-            .set_processed_block_interval(Chain::Solana, 119, 120)
-            .await;
-        assert!(cp.is_none());
-
-        // Exact multiple still works; it crosses from bucket 0 to bucket 1
-        let cp = backlog
-            .set_processed_block_interval(Chain::Solana, 120, 120)
-            .await;
-        assert!(cp.is_some());
-        assert_eq!(cp.unwrap().block_height, 120);
-    }
-
-    #[tokio::test]
-    async fn test_boundary_crossing_first_observed_height_above_interval() {
-        let backlog = Backlog::new();
-        seed_pending_solana_request(&backlog).await;
-
-        // First ever processed block is 500 (prev defaults to 0)
-        // 500 / 120 = 4 > 0, so a boundary was crossed.
-        let cp = backlog
-            .set_processed_block_interval(Chain::Solana, 500, 120)
-            .await
-            .unwrap();
-        assert_eq!(cp.block_height, 500);
-        assert_eq!(cp.pending_requests.len(), 1);
-    }
-
-    #[tokio::test]
-    async fn test_advance_rejects_plain_sign_entries() {
-        let backlog = Backlog::new();
-        let tx = mock_tx(8);
-        let sign_id = tx.sign_id();
-
-        backlog.insert_mock_sign(sign_id, tx.source_chain).await;
-
-        let mut entry = backlog
-            .pending(&tx.source_chain)
-            .read()
-            .await
-            .get(&sign_id)
-            .cloned()
-            .expect("entry should be found");
-        let err = entry
-            .advance(Arc::new(tx))
-            .expect_err("advance should fail for plain Sign requests");
-
-        assert_matches!(err, BacklogError::InvalidAdvanceTransition);
-    }
-
-    #[tokio::test]
-    async fn test_advance_accepts_pending_generation_bidirectional() {
-        let backlog = Backlog::new();
-        let tx = mock_tx(9);
-        let sign_id = tx.sign_id();
-
-        backlog
-            .insert_mock_bidirectional(sign_id, tx.source_chain)
-            .await;
-
-        let entry = backlog
-            .get_by::<Bidirectional<Initial<AnyProgress>>>(tx.source_chain, &sign_id)
-            .await
-            .expect("entry should be found");
-        entry
-            .advance(Arc::new(tx.clone()))
-            .await
-            .expect("advance should accept catchup advancement from PendingGeneration");
-
-        let entry = backlog
-            .get_by::<Bidirectional<Executing>>(tx.source_chain, &sign_id)
-            .await
-            .expect("entry should remain in backlog");
-        assert_eq!(entry.execution_tx().id, tx.id);
-    }
-
-    #[tokio::test]
-    async fn test_advance_accepts_pending_publish_bidirectional() {
-        let backlog = Backlog::new();
-        let tx = mock_tx(10);
-        let sign_id = tx.sign_id();
-
-        let entry = backlog.insert_mock_executing(&tx).await;
-
-        assert_eq!(entry.execution_tx().id, tx.id);
-        let fetched = backlog
-            .get_by::<Bidirectional<Executing>>(tx.source_chain, &sign_id)
-            .await
-            .expect("entry should remain in backlog");
-        assert_eq!(fetched.execution_tx().id, tx.id);
-    }
-
-    #[tokio::test]
-    async fn test_total_pending_increments_on_insert() {
-        let backlog = Backlog::new();
-
-        backlog
-            .insert_mock_sign(SignId::from_u8(1), Chain::Ethereum)
-            .await;
-
-        assert_eq!(backlog.len(), 1);
-        assert!(!backlog.is_empty());
-    }
-
-    #[tokio::test]
-    async fn test_total_pending_ignores_duplicate_inserts() {
-        let backlog = Backlog::new();
-        let sign_id = SignId::from_u8(1);
-
-        // Insert first time
-        backlog.insert_mock_sign(sign_id, Chain::Ethereum).await;
-        assert_eq!(backlog.len(), 1);
-
-        // Insert exactly the same ID again (overwrites)
-        backlog.insert_mock_sign(sign_id, Chain::Ethereum).await;
-        assert_eq!(
-            backlog.len(),
-            1,
-            "Duplicate insert should not increment total"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_total_pending_counts_across_chains() {
-        let backlog = Backlog::new();
-
-        backlog
-            .insert_mock_sign(SignId::from_u8(1), Chain::Ethereum)
-            .await;
-        backlog
-            .insert_mock_sign(SignId::from_u8(2), Chain::Solana)
-            .await;
-
-        assert_eq!(backlog.len(), 2);
-    }
-
-    #[tokio::test]
-    async fn test_total_pending_decrements_on_remove() {
-        let backlog = Backlog::new();
-        let sign_id = SignId::from_u8(1);
-
-        backlog.insert_mock_sign(sign_id, Chain::Ethereum).await;
-        assert_eq!(backlog.len(), 1);
-
-        backlog.remove(Chain::Ethereum, &sign_id).await;
-        assert_eq!(backlog.len(), 0);
-        assert!(backlog.is_empty());
-    }
-
-    #[tokio::test]
-    async fn test_total_pending_ignores_invalid_removes() {
-        let backlog = Backlog::new();
-        let sign_id1 = SignId::from_u8(1);
-        let sign_id2 = SignId::from_u8(2); // Not inserted
-
-        backlog.insert_mock_sign(sign_id1, Chain::Ethereum).await;
-
-        backlog.remove(Chain::Ethereum, &sign_id2).await;
-        assert_eq!(
-            backlog.len(),
-            1,
-            "Removing non-existent ID should not decrement total"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_total_pending_updates_on_clean_recovery() {
-        let backlog = Backlog::new();
-
-        // Populate 3 requests and create a checkpoint
-        for i in 1..=3 {
-            backlog
-                .insert_mock_sign(SignId::from_u8(i), Chain::Ethereum)
-                .await;
-        }
-        backlog.set_processed_block(Chain::Ethereum, 10).await;
-        let checkpoint = backlog.checkpoint(Chain::Ethereum).await.unwrap();
-
-        // Clean backlog recovers the checkpoint
-        let recovered = Backlog::new();
-        assert_eq!(recovered.len(), 0);
-
-        recovered
-            .recover_by_checkpoint(checkpoint)
-            .await
-            .expect("failed to recover");
-
-        assert_eq!(recovered.len(), 3);
-    }
-
-    #[tokio::test]
-    async fn test_total_pending_updates_on_dirty_recovery() {
-        let backlog = Backlog::new();
-
-        // Populate 3 requests and create a checkpoint
-        for i in 1..=3 {
-            backlog
-                .insert_mock_sign(SignId::from_u8(i), Chain::Ethereum)
-                .await;
-        }
-        backlog.set_processed_block(Chain::Ethereum, 10).await;
-        let checkpoint = backlog.checkpoint(Chain::Ethereum).await.unwrap();
-
-        // Dirty backlog has 1 entirely different request before recovery
-        let dirty_backlog = Backlog::new();
-        dirty_backlog
-            .insert_mock_sign(SignId::from_u8(99), Chain::Ethereum)
-            .await;
-
-        assert_eq!(dirty_backlog.len(), 1);
-
-        // Recover from checkpoint (should overwrite the dirty state)
-        dirty_backlog
-            .recover_by_checkpoint(checkpoint)
-            .await
-            .expect("failed to recover");
-
-        assert_eq!(
-            dirty_backlog.len(),
-            3,
-            "Total should reflect exactly the restored checkpoint size, ignoring the overwritten dirty state"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_recovery_keeps_pending_checkpoints() {
-        let backlog = Backlog::new();
-        let chain = Chain::Ethereum;
-        let interval = chain.checkpoint_interval().unwrap();
-
-        backlog.set_processed_block(chain, interval).await.unwrap();
-        backlog
-            .set_processed_block(chain, 2 * interval)
-            .await
-            .unwrap();
-
-        assert_eq!(
-            backlog.pending_checkpoint_count(chain).await,
-            2,
-            "two checkpoints should be pending"
-        );
-
-        // Recover the request backlog without discarding checkpoints that may
-        // still be needed to match an on-chain consensus digest.
-        let recovery_cp = Checkpoint::reset(chain, interval / 2);
-        backlog.recover_by_checkpoint(recovery_cp).await.unwrap();
-        assert_eq!(
-            backlog.pending_checkpoint_count(chain).await,
-            2,
-            "pending checkpoints should remain available for consensus matching"
-        );
-    }
-}
+mod tests;
