+++ chain-signatures/node/src/backlog/request.rs
@@ -594,236 +594,3 @@ impl Backlog {
         SignEntry::generating(request, self)
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::{
-        AnyProgress, Bidirectional, Executing, Final, Generating, Initial, Publishing, Sign,
-    };
-    use crate::backlog::mock::{
-        mock_bidi_request, mock_bidi_response, mock_bidirectional_tx, mock_participants,
-        mock_sign_request, mock_signature_output, BacklogTestExt,
-    };
-    use crate::backlog::{Backlog, BacklogError};
-    use mpc_primitives::{Chain, SignId};
-    use std::sync::Arc;
-
-    #[tokio::test]
-    async fn test_plain_sign_typestate_advance_all_the_way_till_completion() {
-        let backlog = Backlog::new();
-        let sign_id = SignId::from_u8(1);
-        let req = mock_sign_request(sign_id, Chain::Ethereum);
-
-        // 1. Initial entry via Backlog::insert_sign
-        let entry = backlog.insert_sign(Arc::clone(&req)).await;
-        assert_eq!(entry.sign_id(), sign_id);
-        assert_eq!(entry.state(), &Sign(Generating));
-
-        // Verify that querying with the wrong state returns None
-        assert!(backlog
-            .get_by::<Sign<Publishing>>(Chain::Ethereum, &sign_id)
-            .await
-            .is_none());
-
-        // 2. Advance to publishing
-        let (pk, output) = mock_signature_output(&req.args);
-        let pub_entry = entry
-            .advance(pk, &output, mock_participants(), true)
-            .await
-            .expect("advance call should succeed");
-
-        assert_eq!(pub_entry.sign_id(), sign_id);
-        assert_eq!(pub_entry.request().id, sign_id);
-
-        // 3. Verify signature
-        let root_sk = k256::SecretKey::random(&mut rand::thread_rng());
-        let valid_sig = mpc_crypto::generate_signature(&root_sk, &req.args);
-        pub_entry
-            .verify_signature(root_sk.public_key().into(), &valid_sig)
-            .expect("signature should verify");
-
-        // 4. Complete removing from backlog
-        let removed = pub_entry.complete().await;
-        assert!(removed.is_some());
-        assert!(backlog.get(Chain::Ethereum, &sign_id).await.is_none());
-    }
-
-    #[tokio::test]
-    async fn test_plain_sign_entry_chained_advance() {
-        let backlog = Backlog::new();
-        let sign_id = SignId::from_u8(11);
-        let req = mock_sign_request(sign_id, Chain::Ethereum);
-
-        // Chained advance from sign all the way till completion
-        let (pk, output) = mock_signature_output(&req.args);
-        let completed = backlog
-            .insert_sign(req)
-            .await
-            .advance(pk, &output, mock_participants(), true)
-            .await
-            .expect("advance call should succeed")
-            .complete()
-            .await;
-
-        assert!(completed.is_some());
-        assert!(backlog.get(Chain::Ethereum, &sign_id).await.is_none());
-    }
-
-    #[tokio::test]
-    async fn test_bidirectional_typestate_full_lifecycle() {
-        let backlog = Backlog::new();
-        let sign_id = SignId::from_u8(2);
-        let req = mock_bidi_request(sign_id, Chain::Solana);
-
-        // 1. Initial Generating via insert_bidirectional
-        let bidi_entry = backlog.insert_bidirectional(Arc::clone(&req)).await;
-        assert_eq!(bidi_entry.sign_id(), sign_id);
-        assert_eq!(bidi_entry.state(), &Bidirectional(Initial(Generating)));
-
-        // 2. Advance to Publishing
-        let (pk1, output1) = mock_signature_output(&req.args);
-        let pub_entry = bidi_entry
-            .advance(pk1, &output1, mock_participants(), true)
-            .await
-            .expect("advance call should succeed");
-
-        // Verify phase 1 signature
-        let root_sk = k256::SecretKey::random(&mut rand::thread_rng());
-        let sig1 = mpc_crypto::generate_signature(&root_sk, &req.args);
-        pub_entry
-            .verify_signature(root_sk.public_key().into(), &sig1)
-            .expect("sig1 should verify");
-
-        // 3. Advance to Executing
-        let tx = Arc::new(mock_bidirectional_tx(sign_id, Chain::Solana));
-        let exec_entry = pub_entry
-            .advance(Arc::clone(&tx))
-            .await
-            .expect("should advance to executing");
-        assert_eq!(exec_entry.execution_tx().id, tx.id);
-
-        // Verify get_by can retrieve Executing state directly from backlog
-        assert!(backlog
-            .get_by::<Bidirectional<Executing>>(Chain::Solana, &sign_id)
-            .await
-            .is_some());
-        assert!(backlog
-            .get_by::<Bidirectional<Initial<Generating>>>(Chain::Solana, &sign_id)
-            .await
-            .is_none());
-
-        // 4. Advance to Final Generating
-        let response_request = mock_bidi_response(&tx);
-        let final_entry = exec_entry
-            .advance(Arc::clone(&response_request))
-            .await
-            .expect("should advance to final");
-        assert_eq!(final_entry.request().id, sign_id);
-        assert_eq!(
-            final_entry.request().args.payload,
-            response_request.args.payload
-        );
-
-        // Verify get_by can retrieve Final<Generating> state directly from backlog
-        assert!(backlog
-            .get_by::<Bidirectional<Final<Generating>>>(Chain::Solana, &sign_id)
-            .await
-            .is_some());
-
-        // 5. Advance to Final Publishing
-        let (pk2, output2) = mock_signature_output(&response_request.args);
-        let final_pub_entry = final_entry
-            .advance(pk2, &output2, mock_participants(), true)
-            .await
-            .expect("advance call should succeed");
-        assert_eq!(final_pub_entry.respond_request().id, sign_id);
-
-        // Verify phase 2 signature against response_request
-        let sig2 = mpc_crypto::generate_signature(&root_sk, &response_request.args);
-        final_pub_entry
-            .verify_signature(root_sk.public_key().into(), &sig2)
-            .expect("sig2 should verify");
-
-        // 6. Complete removing from backlog
-        let completed = final_pub_entry.complete().await;
-        assert!(completed.is_some());
-        assert!(backlog.get(Chain::Solana, &sign_id).await.is_none());
-    }
-
-    #[tokio::test]
-    async fn test_bidirectional_entry_chained_advances() {
-        let backlog = Backlog::new();
-        let sign_id = SignId::from_u8(22);
-        let req = mock_bidi_request(sign_id, Chain::Solana);
-        let tx = Arc::new(mock_bidirectional_tx(sign_id, Chain::Solana));
-        let response_request = mock_bidi_response(&tx);
-
-        // Start from initial entry, and keep calling advance all the way till completion!
-        let (pk1, output1) = mock_signature_output(&req.args);
-        let (pk2, output2) = mock_signature_output(&response_request.args);
-        let completed = backlog
-            .insert_bidirectional(req)
-            .await
-            .advance(pk1, &output1, mock_participants(), true)
-            .await
-            .expect("advance call should succeed")
-            .advance(tx)
-            .await
-            .expect("advance to executing")
-            .advance(response_request)
-            .await
-            .expect("advance to final generating")
-            .advance(pk2, &output2, mock_participants(), true)
-            .await
-            .expect("advance call should succeed")
-            .complete()
-            .await;
-
-        assert!(completed.is_some());
-        assert!(backlog.get(Chain::Solana, &sign_id).await.is_none());
-    }
-
-    #[tokio::test]
-    async fn test_any_progress_completion() {
-        let backlog = Backlog::new();
-        let sign_id = SignId::from_u8(33);
-
-        let entry = backlog.insert_mock_sign(sign_id, Chain::Ethereum).await;
-        let (pk, output) = mock_signature_output(&entry.request().args);
-        let _ = entry
-            .advance(pk, &output, mock_participants(), true)
-            .await
-            .expect("advance call should succeed");
-
-        // Query with wildcard AnyProgress typestate
-        let entry = backlog
-            .get_by::<Sign<AnyProgress>>(Chain::Ethereum, &sign_id)
-            .await
-            .expect("should match AnyProgress");
-
-        // Complete removing from backlog
-        let completed = entry.complete().await;
-        assert!(completed.is_some());
-        assert!(backlog.get(Chain::Ethereum, &sign_id).await.is_none());
-    }
-
-    #[tokio::test]
-    async fn test_advance_invalid_signature_error() {
-        let backlog = Backlog::new();
-        let sign_id = SignId::from_u8(99);
-        let req = mock_sign_request(sign_id, Chain::Ethereum);
-        let entry = backlog.insert_sign(req).await;
-
-        let wrong_root_sk = k256::SecretKey::random(&mut rand::thread_rng());
-        let wrong_pk = wrong_root_sk.public_key().into();
-        let dummy_output = cait_sith::FullSignature {
-            big_r: k256::AffinePoint::GENERATOR,
-            s: k256::Scalar::ONE,
-        };
-
-        let result = entry
-            .advance(wrong_pk, &dummy_output, mock_participants(), true)
-            .await;
-        assert_eq!(result.unwrap_err(), BacklogError::InvalidSignature);
-    }
-}
